Refactor project to separate components and pages
