+++ news-ai-client/src/pages/About.js
